Modify the mac_versions method so that all versions are retrieved
At the moment, MAC limits version pages to a size of 50, so despite us
setting a limit of 100, only 50 versions are retrieved. This means we
aren't receiving any 5.0.x or 5.1.x versions, and as a result the 5.0
and 5.1 tags aren't being updated. This commit updates the function so
that instead we pull all versions from MAC.