CLIP-1770: Override new Snyk app-vuln scanning default.
Starting January 24th Snyk enabled the --app-vulns flag by default
on snyk container test and snyk container monitor scans. So now it
will look for application vulnerabilities in the container as well
as system vulnerabilities. We are currently not filing tickets for
app vulns as part of our automated scanning. The change brings us
into line with that policy.